Modularized the main file code
