+++ lib/auth_page.dart
@@ -1,0 +1,715 @@
+import 'package:flutter/material.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
+import 'dart:async';
+import 'main.dart';
+// import 'package:flutter_dotenv/flutter_dotenv.dart';
+
+class AuthPage extends StatefulWidget {
+  const AuthPage({super.key});
+
+  @override
+  State<AuthPage> createState() => _AuthPageState();
+}
+
+class _AuthPageState extends State<AuthPage> {
+  bool isLogin = true;
+  final _formKey = GlobalKey<FormState>();
+  final _emailController = TextEditingController();
+  final _passwordController = TextEditingController();
+  final _nameController = TextEditingController();
+  final _courseController = TextEditingController();
+  bool _isPasswordVisible = false;
+  bool _isLoading = false;
+
+  @override
+  void dispose() {
+    _emailController.dispose();
+    _passwordController.dispose();
+    _nameController.dispose();
+    _courseController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.grey[50],
+      appBar: AppBar(
+        backgroundColor: Colors.transparent,
+        elevation: 0,
+        leading: IconButton(
+          icon: const Icon(Icons.arrow_back, color: Colors.black),
+          onPressed: () => Navigator.pop(context),
+        ),
+      ),
+      body: SingleChildScrollView(
+        padding: const EdgeInsets.all(24),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            const SizedBox(height: 20),
+            
+            // Title
+            Text(
+              isLogin ? 'Welcome Back!' : 'Create Account',
+              style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+                fontWeight: FontWeight.bold,
+                color: Colors.black87,
+              ),
+            ),
+            
+            const SizedBox(height: 8),
+            
+            Text(
+              isLogin 
+                ? 'Sign in to continue to College Hub'
+                : 'Join College Hub to get started',
+              style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                color: Colors.grey[600],
+              ),
+            ),
+            
+            const SizedBox(height: 32),
+            
+            // Form
+            Form(
+              key: _formKey,
+              child: Column(
+                children: [
+                  // Name field (only for signup)
+                  if (!isLogin) ...[
+                    TextFormField(
+                      controller: _nameController,
+                      decoration: InputDecoration(
+                        labelText: 'Full Name',
+                        prefixIcon: const Icon(Icons.person),
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        filled: true,
+                        fillColor: Colors.white,
+                      ),
+                      validator: (value) {
+                        if (value == null || value.isEmpty) {
+                          return 'Please enter your name';
+                        }
+                        return null;
+                      },
+                    ),
+                    const SizedBox(height: 16),
+                  ],
+                  
+                  // Email field
+                  TextFormField(
+                    controller: _emailController,
+                    keyboardType: TextInputType.emailAddress,
+                    decoration: InputDecoration(
+                      labelText: 'Email',
+                      prefixIcon: const Icon(Icons.email),
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                      filled: true,
+                      fillColor: Colors.white,
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Please enter your email';
+                      }
+                      if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value)) {
+                        return 'Please enter a valid email';
+                      }
+                      return null;
+                    },
+                  ),
+                  
+                  const SizedBox(height: 16),
+                  
+                  // Course field (only for signup)
+                  if (!isLogin) ...[
+                    TextFormField(
+                      controller: _courseController,
+                      decoration: InputDecoration(
+                        labelText: 'Course/Major',
+                        prefixIcon: const Icon(Icons.school),
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        filled: true,
+                        fillColor: Colors.white,
+                      ),
+                      validator: (value) {
+                        if (value == null || value.isEmpty) {
+                          return 'Please enter your course';
+                        }
+                        return null;
+                      },
+                    ),
+                    const SizedBox(height: 16),
+                  ],
+                  
+                  // Password field
+                  TextFormField(
+                    controller: _passwordController,
+                    obscureText: !_isPasswordVisible,
+                    decoration: InputDecoration(
+                      labelText: 'Password',
+                      prefixIcon: const Icon(Icons.lock),
+                      suffixIcon: IconButton(
+                        icon: Icon(
+                          _isPasswordVisible ? Icons.visibility : Icons.visibility_off,
+                        ),
+                        onPressed: () {
+                          setState(() {
+                            _isPasswordVisible = !_isPasswordVisible;
+                          });
+                        },
+                      ),
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                      filled: true,
+                      fillColor: Colors.white,
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Please enter your password';
+                      }
+                      if (value.length < 6) {
+                        return 'Password must be at least 6 characters';
+                      }
+                      return null;
+                    },
+                  ),
+                  
+                  const SizedBox(height: 24),
+                  
+                  // Submit Button
+                  SizedBox(
+                    width: double.infinity,
+                    height: 56,
+                    child: ElevatedButton(
+                      onPressed: _isLoading ? null : _handleSubmit,
+                      style: ElevatedButton.styleFrom(
+                        backgroundColor: const Color(0xFF6750A4),
+                        foregroundColor: Colors.white,
+                        elevation: 4,
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                      ),
+                      child: _isLoading 
+                        ? const SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: CircularProgressIndicator(
+                              strokeWidth: 2,
+                              valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+                            ),
+                          )
+                        : Text(
+                            isLogin ? 'Sign In' : 'Sign Up',
+                            style: const TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                    ),
+                  ),
+                  
+                  const SizedBox(height: 16),
+                  
+                  // Forgot Password (only for login)
+                  if (isLogin)
+                    TextButton(
+                      onPressed: _handleForgotPassword,
+                      child: const Text('Forgot Password?'),
+                    ),
+                  
+                  const SizedBox(height: 24),
+                  
+                  // Switch between login/signup
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Text(
+                        isLogin ? "Don't have an account? " : "Already have an account? ",
+                        style: TextStyle(color: Colors.grey[600]),
+                      ),
+                      TextButton(
+                        onPressed: () {
+                          setState(() {
+                            isLogin = !isLogin;
+                          });
+                        },
+                        child: Text(
+                          isLogin ? 'Sign Up' : 'Sign In',
+                          style: const TextStyle(
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Future<void> _handleSubmit() async {
+    if (!_formKey.currentState!.validate()) return;
+
+    setState(() {
+      _isLoading = true;
+    });
+
+    try {
+      if (isLogin) {
+        // Sign in existing user
+        final response = await supabase.auth.signInWithPassword(
+          email: _emailController.text.trim(),
+          password: _passwordController.text,
+        );
+
+        if (response.user != null) {
+          // Check if email is confirmed
+          if (response.user!.emailConfirmedAt != null) {
+            _navigateToMainScreen();
+          } else {
+            _showEmailVerificationDialog();
+          }
+        }
+      } else {
+        // Sign up new user
+        final response = await supabase.auth.signUp(
+          email: _emailController.text.trim(),
+          password: _passwordController.text,
+        );
+
+        if (response.user != null) {
+          // Check if email is confirmed (for auto-confirm setups)
+          if (response.user!.emailConfirmedAt != null) {
+            await _createUserProfile(response.user!.id);
+            _navigateToMainScreen();
+          } else {
+            // Show email verification dialog
+            _showEmailVerificationDialog();
+          }
+        }
+      }
+    } on AuthException catch (error) {
+      _showErrorSnackBar(error.message);
+    } catch (error) {
+      _showErrorSnackBar('An unexpected error occurred');
+    } finally {
+      setState(() {
+        _isLoading = false;
+      });
+    }
+  }
+
+  Future<void> _createUserProfile(String userId) async {
+    try {
+      await supabase.from('profiles').insert({
+        'id': userId,
+        'full_name': _nameController.text.trim(),
+        'course': _courseController.text.trim(),
+        'email': _emailController.text.trim(),
+        'created_at': DateTime.now().toIso8601String(),
+      });
+    } catch (error) {
+      // Handle profile creation error
+      print('Error creating profile: $error');
+    }
+  }
+
+  Future<void> _handleForgotPassword() async {
+    if (_emailController.text.isEmpty) {
+      _showErrorSnackBar('Please enter your email address');
+      return;
+    }
+
+    try {
+      await supabase.auth.resetPasswordForEmail(_emailController.text.trim());
+      _showSuccessSnackBar('Password reset email sent!');
+    } on AuthException catch (error) {
+      _showErrorSnackBar(error.message);
+    } catch (error) {
+      _showErrorSnackBar('An unexpected error occurred');
+    }
+  }
+
+  void _navigateToMainScreen() {
+    Navigator.pushReplacement(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const MainScreen(),
+      ),
+    );
+  }
+
+  void _showErrorSnackBar(String message) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: Colors.red,
+      ),
+    );
+  }
+
+  void _showSuccessSnackBar(String message) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: Colors.green,
+      ),
+    );
+  }
+
+  void _showEmailVerificationDialog() {
+    showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (BuildContext context) {
+        return EmailVerificationDialog(
+          email: _emailController.text.trim(),
+          onVerificationComplete: () {
+            // Create profile if it's a new user
+            if (!isLogin) {
+              _createUserProfile(supabase.auth.currentUser!.id);
+            }
+            _navigateToMainScreen();
+          },
+        );
+      },
+    );
+  }
+}
+
+
+
+class EmailVerificationDialog extends StatefulWidget {
+  final String email;
+  final VoidCallback onVerificationComplete;
+
+  const EmailVerificationDialog({
+    Key? key,
+    required this.email,
+    required this.onVerificationComplete,
+  }) : super(key: key);
+
+  @override
+  State<EmailVerificationDialog> createState() => _EmailVerificationDialogState();
+}
+
+class _EmailVerificationDialogState extends State<EmailVerificationDialog> {
+  bool _isResending = false;
+  bool _isChecking = false;
+  Timer? _verificationTimer;
+  int _resendCooldown = 0;
+  Timer? _cooldownTimer;
+
+  @override
+  void initState() {
+    super.initState();
+    _startVerificationCheck();
+  }
+
+  @override
+  void dispose() {
+    _verificationTimer?.cancel();
+    _cooldownTimer?.cancel();
+    super.dispose();
+  }
+
+  void _startVerificationCheck() {
+    _verificationTimer = Timer.periodic(const Duration(seconds: 3), (timer) {
+      _checkEmailVerification();
+    });
+  }
+
+  Future<void> _checkEmailVerification() async {
+    try {
+      setState(() {
+        _isChecking = true;
+      });
+
+      // Refresh the session to get latest user data
+      await supabase.auth.refreshSession();
+      final user = supabase.auth.currentUser;
+      
+      if (user != null && user.emailConfirmedAt != null) {
+        _verificationTimer?.cancel();
+        Navigator.of(context).pop();
+        widget.onVerificationComplete();
+      }
+    } catch (error) {
+      // Handle error silently during background check
+      print('Error checking verification: $error');
+    } finally {
+      setState(() {
+        _isChecking = false;
+      });
+    }
+  }
+
+  Future<void> _resendVerificationEmail() async {
+    if (_resendCooldown > 0) return;
+
+    setState(() {
+      _isResending = true;
+    });
+
+    try {
+      await supabase.auth.resend(
+        type: OtpType.signup,
+        email: widget.email,
+      );
+
+      _showSuccessSnackBar('Verification email sent!');
+      _startResendCooldown();
+    } on AuthException catch (error) {
+      _showErrorSnackBar(error.message);
+    } catch (error) {
+      _showErrorSnackBar('Failed to resend email');
+    } finally {
+      setState(() {
+        _isResending = false;
+      });
+    }
+  }
+
+  void _startResendCooldown() {
+    setState(() {
+      _resendCooldown = 60;
+    });
+
+    _cooldownTimer = Timer.periodic(const Duration(seconds: 1), (timer) {
+      setState(() {
+        _resendCooldown--;
+      });
+
+      if (_resendCooldown <= 0) {
+        timer.cancel();
+      }
+    });
+  }
+
+  void _showErrorSnackBar(String message) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: Colors.red,
+      ),
+    );
+  }
+
+  void _showSuccessSnackBar(String message) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: Colors.green,
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(16),
+      ),
+      title: const Row(
+        children: [
+          Icon(
+            Icons.mail_outline,
+            color: Color(0xFF6750A4),
+            size: 28,
+          ),
+          SizedBox(width: 12),
+          Text(
+            'Verify Your Email',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ],
+      ),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            'We\'ve sent a verification email to:',
+            style: TextStyle(
+              fontSize: 16,
+              color: Colors.grey[600],
+            ),
+          ),
+          const SizedBox(height: 8),
+          Container(
+            padding: const EdgeInsets.all(12),
+            decoration: BoxDecoration(
+              color: Colors.grey[100],
+              borderRadius: BorderRadius.circular(8),
+            ),
+            child: Text(
+              widget.email,
+              style: const TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.w500,
+              ),
+            ),
+          ),
+          const SizedBox(height: 16),
+          Text(
+            'Please check your inbox and click the verification link to continue.',
+            style: TextStyle(
+              fontSize: 14,
+              color: Colors.grey[600],
+            ),
+          ),
+          const SizedBox(height: 16),
+          if (_isChecking)
+            Row(
+              children: [
+                const SizedBox(
+                  width: 16,
+                  height: 16,
+                  child: CircularProgressIndicator(
+                    strokeWidth: 2,
+                  ),
+                ),
+                const SizedBox(width: 12),
+                Text(
+                  'Checking verification...',
+                  style: TextStyle(
+                    fontSize: 14,
+                    color: Colors.grey[600],
+                  ),
+                ),
+              ],
+            ),
+        ],
+      ),
+      actions: [
+        TextButton(
+          onPressed: () {
+            _verificationTimer?.cancel();
+            Navigator.of(context).pop();
+            // Sign out the user since they haven't verified
+            supabase.auth.signOut();
+          },
+          child: const Text('Cancel'),
+        ),
+        ElevatedButton(
+          onPressed: _resendCooldown > 0 || _isResending ? null : _resendVerificationEmail,
+          style: ElevatedButton.styleFrom(
+            backgroundColor: const Color(0xFF6750A4),
+            foregroundColor: Colors.white,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(8),
+            ),
+          ),
+          child: _isResending
+              ? const SizedBox(
+                  width: 16,
+                  height: 16,
+                  child: CircularProgressIndicator(
+                    strokeWidth: 2,
+                    valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+                  ),
+                )
+              : Text(
+                  _resendCooldown > 0 
+                      ? 'Resend (${_resendCooldown}s)'
+                      : 'Resend Email',
+                ),
+        ),
+      ],
+    );
+  }
+}
+
+
+
+class AuthGate extends StatefulWidget {
+  @override
+  _AuthGateState createState() => _AuthGateState();
+}
+
+class _AuthGateState extends State<AuthGate> {
+  @override
+  void initState() {
+    super.initState();
+    _checkInitialSession();
+    _setupAuthListener();
+  }
+
+  // Check if user is already signed in when app starts
+  void _checkInitialSession() {
+    final session = supabase.auth.currentSession;
+    if (session != null && session.user.emailConfirmedAt != null) {
+      // User is already signed in and verified, navigate to main screen
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const MainScreen()),
+        );
+      });
+    } else {
+      // User is not signed in or not verified, navigate to auth page
+      if (session != null && session.user.emailConfirmedAt == null) {
+        // Sign out unverified user
+        supabase.auth.signOut();
+      }
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const AuthPage()),
+        );
+      });
+    }
+  }
+
+  // Listen for auth state changes
+  void _setupAuthListener() {
+    supabase.auth.onAuthStateChange.listen((data) {
+      final session = data.session;
+      if (session != null) {
+        // Check if email is verified
+        if (session.user.emailConfirmedAt != null) {
+          // User is signed in and verified
+          Navigator.pushReplacement(
+            context,
+            MaterialPageRoute(builder: (context) => const MainScreen()),
+          );
+        } else {
+          // User is signed in but not verified, sign them out
+          supabase.auth.signOut();
+        }
+      } else {
+        // User is signed out
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const AuthPage()),
+        );
+      }
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
+
