+++ lib/main.dart
@@ -2,6 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'dart:async';
+import 'landing_page.dart';
+import 'profile_screen.dart';
+import 'opportunity_screen.dart';
+import 'notes_screen.dart';
 
 
 //Main Function
@@ -24,9 +28,7 @@ await Supabase.initialize(
   runApp(MyApp());
 }
 
-
 final supabase = Supabase.instance.client;
-
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
@@ -48,852 +50,7 @@ class MyApp extends StatelessWidget {
   }
 }
 
-
-//Landing Page
-class LandingPage extends StatelessWidget {
-  const LandingPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        width: double.infinity,
-        decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topLeft,
-            end: Alignment.bottomRight,
-            colors: [
-              Color(0xFF667eea),
-              Color(0xFF764ba2),
-            ],
-          ),
-        ),
-        child: SafeArea(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 24.0),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const Spacer(),
-                
-                // App Logo/Icon
-                Container(
-                  width: 120,
-                  height: 120,
-                  decoration: BoxDecoration(
-                    color: Colors.white.withOpacity(0.2),
-                    borderRadius: BorderRadius.circular(30),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black.withOpacity(0.1),
-                        blurRadius: 20,
-                        offset: const Offset(0, 8),
-                      ),
-                    ],
-                  ),
-                  child: const Icon(
-                    Icons.school,
-                    size: 60,
-                    color: Colors.white,
-                  ),
-                ),
-                
-                const SizedBox(height: 32),
-                
-                // App Title
-                Text(
-                  'College Hub',
-                  style: Theme.of(context).textTheme.headlineLarge?.copyWith(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 36,
-                  ),
-                ),
-                
-                const SizedBox(height: 16),
-                
-                // Subtitle
-                Text(
-                  'Share knowledge, discover opportunities',
-                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                    color: Colors.white.withOpacity(0.9),
-                    fontSize: 18,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-                
-                const SizedBox(height: 8),
-                
-                // Description
-                Text(
-                  'Connect with fellow students, share study notes,\nand find amazing opportunities',
-                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Colors.white.withOpacity(0.8),
-                    fontSize: 16,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-                
-                const Spacer(),
-                
-                // Get Started Button
-                SizedBox(
-                  width: double.infinity,
-                  height: 56,
-                  child: ElevatedButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => const AuthPage(),
-                        ),
-                      );
-                    },
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.white,
-                      foregroundColor: const Color(0xFF764ba2),
-                      elevation: 8,
-                      shadowColor: Colors.black.withOpacity(0.3),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(16),
-                      ),
-                    ),
-                    child: const Text(
-                      'Get Started',
-                      style: TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-                ),
-                
-                const SizedBox(height: 48),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-
-
-// Authentication Page (Login/Signup)
-class AuthPage extends StatefulWidget {
-  const AuthPage({super.key});
-
-  @override
-  State<AuthPage> createState() => _AuthPageState();
-}
-
-class _AuthPageState extends State<AuthPage> {
-  bool isLogin = true;
-  final _formKey = GlobalKey<FormState>();
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
-  final _nameController = TextEditingController();
-  final _courseController = TextEditingController();
-  bool _isPasswordVisible = false;
-  bool _isLoading = false;
-
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    _nameController.dispose();
-    _courseController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.grey[50],
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back, color: Colors.black),
-          onPressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: SingleChildScrollView(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            const SizedBox(height: 20),
-            
-            // Title
-            Text(
-              isLogin ? 'Welcome Back!' : 'Create Account',
-              style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-                fontWeight: FontWeight.bold,
-                color: Colors.black87,
-              ),
-            ),
-            
-            const SizedBox(height: 8),
-            
-            Text(
-              isLogin 
-                ? 'Sign in to continue to College Hub'
-                : 'Join College Hub to get started',
-              style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                color: Colors.grey[600],
-              ),
-            ),
-            
-            const SizedBox(height: 32),
-            
-            // Form
-            Form(
-              key: _formKey,
-              child: Column(
-                children: [
-                  // Name field (only for signup)
-                  if (!isLogin) ...[
-                    TextFormField(
-                      controller: _nameController,
-                      decoration: InputDecoration(
-                        labelText: 'Full Name',
-                        prefixIcon: const Icon(Icons.person),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.white,
-                      ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Please enter your name';
-                        }
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 16),
-                  ],
-                  
-                  // Email field
-                  TextFormField(
-                    controller: _emailController,
-                    keyboardType: TextInputType.emailAddress,
-                    decoration: InputDecoration(
-                      labelText: 'Email',
-                      prefixIcon: const Icon(Icons.email),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                      filled: true,
-                      fillColor: Colors.white,
-                    ),
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter your email';
-                      }
-                      if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value)) {
-                        return 'Please enter a valid email';
-                      }
-                      return null;
-                    },
-                  ),
-                  
-                  const SizedBox(height: 16),
-                  
-                  // Course field (only for signup)
-                  if (!isLogin) ...[
-                    TextFormField(
-                      controller: _courseController,
-                      decoration: InputDecoration(
-                        labelText: 'Course/Major',
-                        prefixIcon: const Icon(Icons.school),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.white,
-                      ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Please enter your course';
-                        }
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 16),
-                  ],
-                  
-                  // Password field
-                  TextFormField(
-                    controller: _passwordController,
-                    obscureText: !_isPasswordVisible,
-                    decoration: InputDecoration(
-                      labelText: 'Password',
-                      prefixIcon: const Icon(Icons.lock),
-                      suffixIcon: IconButton(
-                        icon: Icon(
-                          _isPasswordVisible ? Icons.visibility : Icons.visibility_off,
-                        ),
-                        onPressed: () {
-                          setState(() {
-                            _isPasswordVisible = !_isPasswordVisible;
-                          });
-                        },
-                      ),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                      filled: true,
-                      fillColor: Colors.white,
-                    ),
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter your password';
-                      }
-                      if (value.length < 6) {
-                        return 'Password must be at least 6 characters';
-                      }
-                      return null;
-                    },
-                  ),
-                  
-                  const SizedBox(height: 24),
-                  
-                  // Submit Button
-                  SizedBox(
-                    width: double.infinity,
-                    height: 56,
-                    child: ElevatedButton(
-                      onPressed: _isLoading ? null : _handleSubmit,
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: const Color(0xFF6750A4),
-                        foregroundColor: Colors.white,
-                        elevation: 4,
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                      ),
-                      child: _isLoading 
-                        ? const SizedBox(
-                            height: 20,
-                            width: 20,
-                            child: CircularProgressIndicator(
-                              strokeWidth: 2,
-                              valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                            ),
-                          )
-                        : Text(
-                            isLogin ? 'Sign In' : 'Sign Up',
-                            style: const TextStyle(
-                              fontSize: 18,
-                              fontWeight: FontWeight.bold,
-                            ),
-                          ),
-                    ),
-                  ),
-                  
-                  const SizedBox(height: 16),
-                  
-                  // Forgot Password (only for login)
-                  if (isLogin)
-                    TextButton(
-                      onPressed: _handleForgotPassword,
-                      child: const Text('Forgot Password?'),
-                    ),
-                  
-                  const SizedBox(height: 24),
-                  
-                  // Switch between login/signup
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Text(
-                        isLogin ? "Don't have an account? " : "Already have an account? ",
-                        style: TextStyle(color: Colors.grey[600]),
-                      ),
-                      TextButton(
-                        onPressed: () {
-                          setState(() {
-                            isLogin = !isLogin;
-                          });
-                        },
-                        child: Text(
-                          isLogin ? 'Sign Up' : 'Sign In',
-                          style: const TextStyle(
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Future<void> _handleSubmit() async {
-    if (!_formKey.currentState!.validate()) return;
-
-    setState(() {
-      _isLoading = true;
-    });
-
-    try {
-      if (isLogin) {
-        // Sign in existing user
-        final response = await supabase.auth.signInWithPassword(
-          email: _emailController.text.trim(),
-          password: _passwordController.text,
-        );
-
-        if (response.user != null) {
-          // Check if email is confirmed
-          if (response.user!.emailConfirmedAt != null) {
-            _navigateToMainScreen();
-          } else {
-            _showEmailVerificationDialog();
-          }
-        }
-      } else {
-        // Sign up new user
-        final response = await supabase.auth.signUp(
-          email: _emailController.text.trim(),
-          password: _passwordController.text,
-        );
-
-        if (response.user != null) {
-          // Check if email is confirmed (for auto-confirm setups)
-          if (response.user!.emailConfirmedAt != null) {
-            await _createUserProfile(response.user!.id);
-            _navigateToMainScreen();
-          } else {
-            // Show email verification dialog
-            _showEmailVerificationDialog();
-          }
-        }
-      }
-    } on AuthException catch (error) {
-      _showErrorSnackBar(error.message);
-    } catch (error) {
-      _showErrorSnackBar('An unexpected error occurred');
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  Future<void> _createUserProfile(String userId) async {
-    try {
-      await supabase.from('profiles').insert({
-        'id': userId,
-        'full_name': _nameController.text.trim(),
-        'course': _courseController.text.trim(),
-        'email': _emailController.text.trim(),
-        'created_at': DateTime.now().toIso8601String(),
-      });
-    } catch (error) {
-      // Handle profile creation error
-      print('Error creating profile: $error');
-    }
-  }
-
-  Future<void> _handleForgotPassword() async {
-    if (_emailController.text.isEmpty) {
-      _showErrorSnackBar('Please enter your email address');
-      return;
-    }
-
-    try {
-      await supabase.auth.resetPasswordForEmail(_emailController.text.trim());
-      _showSuccessSnackBar('Password reset email sent!');
-    } on AuthException catch (error) {
-      _showErrorSnackBar(error.message);
-    } catch (error) {
-      _showErrorSnackBar('An unexpected error occurred');
-    }
-  }
-
-  void _navigateToMainScreen() {
-    Navigator.pushReplacement(
-      context,
-      MaterialPageRoute(
-        builder: (context) => const MainScreen(),
-      ),
-    );
-  }
-
-  void _showErrorSnackBar(String message) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: Colors.red,
-      ),
-    );
-  }
-
-  void _showSuccessSnackBar(String message) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: Colors.green,
-      ),
-    );
-  }
-
-  void _showEmailVerificationDialog() {
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return EmailVerificationDialog(
-          email: _emailController.text.trim(),
-          onVerificationComplete: () {
-            // Create profile if it's a new user
-            if (!isLogin) {
-              _createUserProfile(supabase.auth.currentUser!.id);
-            }
-            _navigateToMainScreen();
-          },
-        );
-      },
-    );
-  }
-}
-
-
-
-class EmailVerificationDialog extends StatefulWidget {
-  final String email;
-  final VoidCallback onVerificationComplete;
-
-  const EmailVerificationDialog({
-    Key? key,
-    required this.email,
-    required this.onVerificationComplete,
-  }) : super(key: key);
-
-  @override
-  State<EmailVerificationDialog> createState() => _EmailVerificationDialogState();
-}
-
-class _EmailVerificationDialogState extends State<EmailVerificationDialog> {
-  bool _isResending = false;
-  bool _isChecking = false;
-  Timer? _verificationTimer;
-  int _resendCooldown = 0;
-  Timer? _cooldownTimer;
-
-  @override
-  void initState() {
-    super.initState();
-    _startVerificationCheck();
-  }
-
-  @override
-  void dispose() {
-    _verificationTimer?.cancel();
-    _cooldownTimer?.cancel();
-    super.dispose();
-  }
-
-  void _startVerificationCheck() {
-    _verificationTimer = Timer.periodic(const Duration(seconds: 3), (timer) {
-      _checkEmailVerification();
-    });
-  }
-
-  Future<void> _checkEmailVerification() async {
-    try {
-      setState(() {
-        _isChecking = true;
-      });
-
-      // Refresh the session to get latest user data
-      await supabase.auth.refreshSession();
-      final user = supabase.auth.currentUser;
-      
-      if (user != null && user.emailConfirmedAt != null) {
-        _verificationTimer?.cancel();
-        Navigator.of(context).pop();
-        widget.onVerificationComplete();
-      }
-    } catch (error) {
-      // Handle error silently during background check
-      print('Error checking verification: $error');
-    } finally {
-      setState(() {
-        _isChecking = false;
-      });
-    }
-  }
-
-  Future<void> _resendVerificationEmail() async {
-    if (_resendCooldown > 0) return;
-
-    setState(() {
-      _isResending = true;
-    });
-
-    try {
-      await supabase.auth.resend(
-        type: OtpType.signup,
-        email: widget.email,
-      );
-
-      _showSuccessSnackBar('Verification email sent!');
-      _startResendCooldown();
-    } on AuthException catch (error) {
-      _showErrorSnackBar(error.message);
-    } catch (error) {
-      _showErrorSnackBar('Failed to resend email');
-    } finally {
-      setState(() {
-        _isResending = false;
-      });
-    }
-  }
-
-  void _startResendCooldown() {
-    setState(() {
-      _resendCooldown = 60;
-    });
-
-    _cooldownTimer = Timer.periodic(const Duration(seconds: 1), (timer) {
-      setState(() {
-        _resendCooldown--;
-      });
-
-      if (_resendCooldown <= 0) {
-        timer.cancel();
-      }
-    });
-  }
-
-  void _showErrorSnackBar(String message) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: Colors.red,
-      ),
-    );
-  }
-
-  void _showSuccessSnackBar(String message) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: Colors.green,
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(16),
-      ),
-      title: const Row(
-        children: [
-          Icon(
-            Icons.mail_outline,
-            color: Color(0xFF6750A4),
-            size: 28,
-          ),
-          SizedBox(width: 12),
-          Text(
-            'Verify Your Email',
-            style: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        ],
-      ),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'We\'ve sent a verification email to:',
-            style: TextStyle(
-              fontSize: 16,
-              color: Colors.grey[600],
-            ),
-          ),
-          const SizedBox(height: 8),
-          Container(
-            padding: const EdgeInsets.all(12),
-            decoration: BoxDecoration(
-              color: Colors.grey[100],
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Text(
-              widget.email,
-              style: const TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ),
-          const SizedBox(height: 16),
-          Text(
-            'Please check your inbox and click the verification link to continue.',
-            style: TextStyle(
-              fontSize: 14,
-              color: Colors.grey[600],
-            ),
-          ),
-          const SizedBox(height: 16),
-          if (_isChecking)
-            Row(
-              children: [
-                const SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                  ),
-                ),
-                const SizedBox(width: 12),
-                Text(
-                  'Checking verification...',
-                  style: TextStyle(
-                    fontSize: 14,
-                    color: Colors.grey[600],
-                  ),
-                ),
-              ],
-            ),
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: () {
-            _verificationTimer?.cancel();
-            Navigator.of(context).pop();
-            // Sign out the user since they haven't verified
-            supabase.auth.signOut();
-          },
-          child: const Text('Cancel'),
-        ),
-        ElevatedButton(
-          onPressed: _resendCooldown > 0 || _isResending ? null : _resendVerificationEmail,
-          style: ElevatedButton.styleFrom(
-            backgroundColor: const Color(0xFF6750A4),
-            foregroundColor: Colors.white,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(8),
-            ),
-          ),
-          child: _isResending
-              ? const SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                    valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                  ),
-                )
-              : Text(
-                  _resendCooldown > 0 
-                      ? 'Resend (${_resendCooldown}s)'
-                      : 'Resend Email',
-                ),
-        ),
-      ],
-    );
-  }
-}
-
-
-
-class AuthGate extends StatefulWidget {
-  @override
-  _AuthGateState createState() => _AuthGateState();
-}
-
-class _AuthGateState extends State<AuthGate> {
-  @override
-  void initState() {
-    super.initState();
-    _checkInitialSession();
-    _setupAuthListener();
-  }
-
-  // Check if user is already signed in when app starts
-  void _checkInitialSession() {
-    final session = supabase.auth.currentSession;
-    if (session != null && session.user.emailConfirmedAt != null) {
-      // User is already signed in and verified, navigate to main screen
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const MainScreen()),
-        );
-      });
-    } else {
-      // User is not signed in or not verified, navigate to auth page
-      if (session != null && session.user.emailConfirmedAt == null) {
-        // Sign out unverified user
-        supabase.auth.signOut();
-      }
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const AuthPage()),
-        );
-      });
-    }
-  }
-
-  // Listen for auth state changes
-  void _setupAuthListener() {
-    supabase.auth.onAuthStateChange.listen((data) {
-      final session = data.session;
-      if (session != null) {
-        // Check if email is verified
-        if (session.user.emailConfirmedAt != null) {
-          // User is signed in and verified
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const MainScreen()),
-          );
-        } else {
-          // User is signed in but not verified, sign them out
-          supabase.auth.signOut();
-        }
-      } else {
-        // User is signed out
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const AuthPage()),
-        );
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: CircularProgressIndicator(),
-      ),
-    );
-  }
-}
-
-
-
 //Main Screen
-
 class MainScreen extends StatefulWidget {
   const MainScreen({super.key});
 
@@ -1157,842 +314,3 @@ class HomeScreen extends StatelessWidget {
     );
   }
 }
-
-class NotesScreen extends StatefulWidget {
-  const NotesScreen({super.key});
-
-  @override
-  State createState() => _NotesScreenState();
-}
-
-class _NotesScreenState extends State {
-  String selectedCategory = 'All';
-  final List categories = ['All', 'Computer Science', 'Mathematics', 'Physics', 'Chemistry', 'Biology'];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Notes'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.search),
-            onPressed: () {},
-          ),
-          IconButton(
-            icon: const Icon(Icons.filter_list),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: Column(
-        children: [
-          // Category Filter
-          Container(
-            height: 60,
-            padding: const EdgeInsets.symmetric(vertical: 8),
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              itemCount: categories.length,
-              itemBuilder: (context, index) {
-                return Padding(
-                  padding: EdgeInsets.only(
-                    left: index == 0 ? 16 : 8,
-                    right: index == categories.length - 1 ? 16 : 0,
-                  ),
-                  child: FilterChip(
-                    label: Text(categories[index]),
-                    selected: selectedCategory == categories[index],
-                    onSelected: (selected) {
-                      setState(() {
-                        selectedCategory = categories[index];
-                      });
-                    },
-                  ),
-                );
-              },
-            ),
-          ),
-
-          // Notes List
-          Expanded(
-            child: ListView.builder(
-              padding: const EdgeInsets.all(16),
-              itemCount: 10,
-              itemBuilder: (context, index) {
-                return _buildNoteCard(context, index);
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildNoteCard(BuildContext context, int index) {
-    final notes = [
-      {
-        'title': 'Advanced Data Structures',
-        'subject': 'Computer Science',
-        'author': 'Sarah Wilson',
-        'rating': 4.8,
-        'downloads': 234,
-        'date': '2 days ago'
-      },
-      {
-        'title': 'Calculus II - Integration Techniques',
-        'subject': 'Mathematics',
-        'author': 'Robert Chen',
-        'rating': 4.6,
-        'downloads': 156,
-        'date': '5 days ago'
-      },
-      {
-        'title': 'Quantum Mechanics Basics',
-        'subject': 'Physics',
-        'author': 'Emily Davis',
-        'rating': 4.9,
-        'downloads': 89,
-        'date': '1 week ago'
-      },
-    ];
-
-    final note = notes[index % notes.length];
-
-    return Card(
-      margin: const EdgeInsets.only(bottom: 16),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Expanded(
-                  child: Text(
-                    note['title'] as String,
-                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.more_vert),
-                  onPressed: () {},
-                ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            Row(
-              children: [
-                Chip(
-                  label: Text(
-                    note['subject'] as String,
-                    style: const TextStyle(fontSize: 12),
-                  ),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                const SizedBox(width: 8),
-                Text(
-                  'by ${note['author']}',
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Icon(Icons.star, size: 16, color: Colors.amber),
-                const SizedBox(width: 4),
-                Text('${note['rating']}'),
-                const SizedBox(width: 16),
-                Icon(Icons.download, size: 16, color: Colors.grey[600]),
-                const SizedBox(width: 4),
-                Text('${note['downloads']}'),
-                const Spacer(),
-                Text(
-                  note['date'] as String,
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.visibility),
-                    label: const Text('Preview'),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: FilledButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.download),
-                    label: const Text('Download'),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class OpportunitiesScreen extends StatefulWidget {
-  const OpportunitiesScreen({super.key});
-
-  @override
-  State createState() => _OpportunitiesScreenState();
-}
-
-class _OpportunitiesScreenState extends State {
-  String selectedType = 'All';
-  final List types = ['All', 'Internship', 'Job', 'Research', 'Event', 'Workshop'];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Opportunities'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.search),
-            onPressed: () {},
-          ),
-          IconButton(
-            icon: const Icon(Icons.filter_list),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: Column(
-        children: [
-          // Type Filter
-          Container(
-            height: 60,
-            padding: const EdgeInsets.symmetric(vertical: 8),
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              itemCount: types.length,
-              itemBuilder: (context, index) {
-                return Padding(
-                  padding: EdgeInsets.only(
-                    left: index == 0 ? 16 : 8,
-                    right: index == types.length - 1 ? 16 : 0,
-                  ),
-                  child: FilterChip(
-                    label: Text(types[index]),
-                    selected: selectedType == types[index],
-                    onSelected: (selected) {
-                      setState(() {
-                        selectedType = types[index];
-                      });
-                    },
-                  ),
-                );
-              },
-            ),
-          ),
-
-          // Opportunities List
-          Expanded(
-            child: ListView.builder(
-              padding: const EdgeInsets.all(16),
-              itemCount: 8,
-              itemBuilder: (context, index) {
-                return _buildOpportunityCard(context, index);
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildOpportunityCard(BuildContext context, int index) {
-    final opportunities = [
-      {
-        'title': 'Software Engineering Internship',
-        'company': 'Tech Innovators Inc.',
-        'type': 'Internship',
-        'location': 'San Francisco, CA',
-        'deadline': '3 days left',
-        'salary': '\$3,000/month',
-        'urgent': true,
-      },
-      {
-        'title': 'Data Science Research Assistant',
-        'company': 'University Research Lab',
-        'type': 'Research',
-        'location': 'Cambridge, MA',
-        'deadline': '2 weeks left',
-        'salary': '\$2,500/month',
-        'urgent': false,
-      },
-      {
-        'title': 'Machine Learning Workshop',
-        'company': 'AI Academy',
-        'type': 'Workshop',
-        'location': 'Online',
-        'deadline': '5 days left',
-        'salary': 'Free',
-        'urgent': false,
-      },
-    ];
-
-    final opportunity = opportunities[index % opportunities.length];
-
-    return Card(
-      margin: const EdgeInsets.only(bottom: 16),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        opportunity['title'] as String,
-                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 4),
-                      Text(
-                        opportunity['company'] as String,
-                        style: Theme.of(context).textTheme.bodyMedium,
-                      ),
-                    ],
-                  ),
-                ),
-                if (opportunity['urgent'] as bool)
-                  Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    decoration: BoxDecoration(
-                      color: Colors.red,
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    child: const Text(
-                      'URGENT',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 10,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Chip(
-                  label: Text(
-                    opportunity['type'] as String,
-                    style: const TextStyle(fontSize: 12),
-                  ),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                const SizedBox(width: 8),
-                Icon(Icons.location_on, size: 16, color: Colors.grey[600]),
-                const SizedBox(width: 4),
-                Text(
-                  opportunity['location'] as String,
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Icon(Icons.attach_money, size: 16, color: Colors.green),
-                const SizedBox(width: 4),
-                Text(
-                  opportunity['salary'] as String,
-                  style: const TextStyle(
-                    fontWeight: FontWeight.bold,
-                    color: Colors.green,
-                  ),
-                ),
-                const Spacer(),
-                Icon(Icons.access_time, size: 16, color: Colors.orange),
-                const SizedBox(width: 4),
-                Text(
-                  opportunity['deadline'] as String,
-                  style: const TextStyle(
-                    fontWeight: FontWeight.bold,
-                    color: Colors.orange,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.info),
-                    label: const Text('Details'),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: FilledButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.send),
-                    label: const Text('Apply'),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-
-
-//Profile Screen 
-class ProfileScreen extends StatefulWidget {
-  const ProfileScreen({Key? key}) : super(key: key);
-
-  @override
-  State<ProfileScreen> createState() => _ProfileScreenState();
-}
-
-class _ProfileScreenState extends State<ProfileScreen> {
-  bool _isLoading = true;
-  Map<String, dynamic>? _userProfile;
-  String? _error;
-
-
-  @override
-  void initState() {
-    super.initState();
-    _loadUserProfile();
-  }
-
-  Future<void> _loadUserProfile() async {
-    if (!mounted) return;
-    
-    setState(() {
-      _isLoading = true;
-      _error = null;
-    });
-
-    try {
-      final user = supabase.auth.currentUser;
-      if (user == null) {
-        setState(() {
-          _error = 'User not authenticated';
-          _isLoading = false;
-        });
-        return;
-      }
-
-
-      // Try to get existing profile
-      final response = await supabase
-          .from('profiles')
-          .select('*')
-          .eq('id', user.id)
-          .maybeSingle();
-
-
-      if (mounted) {
-        if (response == null) {
-          print('No profile found, creating new one'); // Debug log
-          
-          // No profile found, create one
-          final newProfile = {
-            'id': user.id,
-            'email': user.email ?? 'No email',
-            'full_name': user.userMetadata?['full_name'] ?? 
-                        user.userMetadata?['name'] ?? 
-                        'Unknown User',
-            'course': null,
-          };
-
-          try {
-            // Insert new profile into database
-            final insertedProfile = await supabase
-                .from('profiles')
-                .insert(newProfile)
-                .select()
-                .single();
-            
-            print('Profile created: $insertedProfile'); // Debug log
-            
-            setState(() {
-              _userProfile = insertedProfile;
-              _isLoading = false;
-            });
-          } catch (insertError) {
-            print('Insert error: $insertError'); // Debug log
-            
-            // If insert fails, use temporary profile
-            setState(() {
-              _userProfile = newProfile;
-              _isLoading = false;
-            });
-          }
-        } else {
-          
-          setState(() {
-            _userProfile = response;
-            _isLoading = false;
-          });
-        }
-      }
-    } catch (error) {
-      print('Load profile error: $error'); // Debug log
-      
-      if (mounted) {
-        setState(() {
-          _error = 'Failed to load profile: ${error.toString()}';
-          _isLoading = false;
-        });
-      }
-    }
-  }
-
-  Future<void> _signOut() async {
-    // Show confirmation dialog
-    final shouldSignOut = await showDialog<bool>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(16),
-          ),
-          title: const Text('Sign Out'),
-          content: const Text('Are you sure you want to sign out?'),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(false),
-              child: const Text('Cancel'),
-            ),
-            ElevatedButton(
-              onPressed: () => Navigator.of(context).pop(true),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.red,
-                foregroundColor: Colors.white,
-              ),
-              child: const Text('Sign Out'),
-            ),
-          ],
-        );
-      },
-    );
-
-    if (shouldSignOut == true) {
-      try {
-        await supabase.auth.signOut();
-        
-        if (mounted) {
-          // Show success message
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('Signed out successfully'),
-              backgroundColor: Colors.green,
-              duration: Duration(seconds: 1),
-            ),
-          );
-
-          // Navigate to landing page - replace with your actual landing page route
-          // Navigator.of(context).pushNamedAndRemoveUntil(
-          //   '/', // Replace with your landing page route name
-          //   (route) => false, // Remove all previous routes
-          // );
-          
-          // Alternative: If you're using a specific landing page widget
-          Navigator.of(context).pushAndRemoveUntil(
-            MaterialPageRoute(builder: (context) => const LandingPage()),
-            (route) => false,
-          );
-        }
-      } catch (error) {
-        if (mounted) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              content: Text('Error signing out: $error'),
-              backgroundColor: Colors.red,
-            ),
-          );
-        }
-      }
-    }
-  }
-
-  String _getInitials(String? name) {
-    if (name == null || name.isEmpty) {
-      // Try to get name from user metadata if profile name is empty
-      final user = supabase.auth.currentUser;
-      final metadataName = user?.userMetadata?['full_name']?.toString() ?? 
-                          user?.userMetadata?['name']?.toString() ?? 
-                          user?.email?.split('@')[0] ?? 'U';
-      name = metadataName;
-    }
-    
-    if (name.isEmpty) return 'U';
-    
-    final parts = name.trim().split(' ');
-    if (parts.length >= 2) {
-      return '${parts[0][0]}${parts[1][0]}'.toUpperCase();
-    } else {
-      return name[0].toUpperCase();
-    }
-  }
-
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Profile'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.edit),
-            onPressed: () {
-              // TODO: Navigate to edit profile page
-              ScaffoldMessenger.of(context).showSnackBar(
-                const SnackBar(
-                  content: Text('Edit profile feature coming soon!'),
-                ),
-              );
-            },
-          ),
-        ],
-      ),
-      body: _isLoading
-          ? const Center(child: CircularProgressIndicator())
-          : _error != null
-              ? Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Icon(
-                        Icons.error_outline,
-                        size: 64,
-                        color: Colors.red[300],
-                      ),
-                      const SizedBox(height: 16),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 16),
-                        child: Text(
-                          _error!,
-                          style: Theme.of(context).textTheme.titleMedium,
-                          textAlign: TextAlign.center,
-                        ),
-                      ),
-                      const SizedBox(height: 16),
-                      ElevatedButton(
-                        onPressed: _loadUserProfile,
-                        child: const Text('Retry'),
-                      ),
-                    ],
-                  ),
-                )
-              : RefreshIndicator(
-                  onRefresh: _loadUserProfile,
-                  child: SingleChildScrollView(
-                    physics: const AlwaysScrollableScrollPhysics(),
-                    padding: const EdgeInsets.all(16),
-                    child: Column(
-                      children: [
-                        // Profile Header
-                        Card(
-                          child: Padding(
-                            padding: const EdgeInsets.all(20),
-                            child: Column(
-                              children: [
-                                CircleAvatar(
-                                  radius: 50,
-                                  backgroundColor: Theme.of(context).colorScheme.primary,
-                                  child: Text(
-                                    _getInitials(_userProfile?['full_name']),
-                                    style: const TextStyle(
-                                      fontSize: 32,
-                                      fontWeight: FontWeight.bold,
-                                      color: Colors.white,
-                                    ),
-                                  ),
-                                ),
-                                const SizedBox(height: 16),
-                                Text(
-                                  _userProfile?['full_name']?.toString().isNotEmpty == true 
-                                    ? _userProfile!['full_name'].toString()
-                                    : supabase.auth.currentUser?.userMetadata?['full_name']?.toString() 
-                                    ?? supabase.auth.currentUser?.userMetadata?['name']?.toString()
-                                    ?? 'Unknown User',
-                                  style: Theme.of(context).textTheme.headlineSmall,
-                                ),
-                                const SizedBox(height: 8),
-                                Text(
-                                  _userProfile?['course']?.toString().isNotEmpty == true 
-                                    ? _userProfile!['course'].toString()
-                                    : 'No Course Selected',
-                                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                                    fontWeight: FontWeight.w500,
-                                    color: _userProfile?['course']?.toString().isNotEmpty == true 
-                                      ? Theme.of(context).colorScheme.primary
-                                      : Colors.grey[600],
-                                  ),
-                                ),
-                                const SizedBox(height: 4),
-                                Text(
-                                  _userProfile?['email']?.toString() ?? 
-                                  supabase.auth.currentUser?.email ?? 'No Email',
-                                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                                    color: Colors.grey[600],
-                                  ),
-                                ),
-                                const SizedBox(height: 8),
-                                Container(
-                                  padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                                  decoration: BoxDecoration(
-                                    color: Colors.green.withOpacity(0.1),
-                                    borderRadius: BorderRadius.circular(12),
-                                  ),
-                                  child: Text(
-                                    supabase.auth.currentUser?.emailConfirmedAt != null 
-                                        ? 'Verified' 
-                                        : 'Unverified',
-                                    style: TextStyle(
-                                      color: Colors.green[700],
-                                      fontSize: 12,
-                                      fontWeight: FontWeight.w500,
-                                    ),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 20),
-
-                        // Stats Cards
-                        Row(
-                          children: [
-                            Expanded(
-                              child: _buildStatCard(context, 'Notes Shared', '12', Icons.note),
-                            ),
-                            const SizedBox(width: 16),
-                            Expanded(
-                              child: _buildStatCard(context, 'Downloads', '456', Icons.download),
-                            ),
-                          ],
-                        ),
-                        const SizedBox(height: 16),
-                        Row(
-                          children: [
-                            Expanded(
-                              child: _buildStatCard(context, 'Applications', '8', Icons.send),
-                            ),
-                            const SizedBox(width: 16),
-                            Expanded(
-                              child: _buildStatCard(context, 'Bookmarks', '34', Icons.bookmark),
-                            ),
-                          ],
-                        ),
-                        const SizedBox(height: 20),
-
-                        // Menu Items
-                        Card(
-                          child: Column(
-                            children: [
-                              _buildMenuItem(context, 'My Notes', Icons.note, () {
-                                // TODO: Navigate to notes page
-                              }),
-                              _buildMenuItem(context, 'My Applications', Icons.work, () {
-                                // TODO: Navigate to applications page
-                              }),
-                              _buildMenuItem(context, 'Bookmarks', Icons.bookmark, () {
-                                // TODO: Navigate to bookmarks page
-                              }),
-                              _buildMenuItem(context, 'Settings', Icons.settings, () {
-                                // TODO: Navigate to settings page
-                              }),
-                              _buildMenuItem(context, 'Help & Support', Icons.help, () {
-                                // TODO: Navigate to help page
-                              }),
-                              _buildMenuItem(context, 'About', Icons.info, () {
-                                // TODO: Show about dialog
-                              }),
-                            ],
-                          ),
-                        ),
-                        const SizedBox(height: 20),
-
-                        // Logout Button
-                        SizedBox(
-                          width: double.infinity,
-                          child: OutlinedButton.icon(
-                            onPressed: _signOut,
-                            icon: const Icon(Icons.logout),
-                            label: const Text('Logout'),
-                            style: OutlinedButton.styleFrom(
-                              foregroundColor: Colors.red,
-                              side: const BorderSide(color: Colors.red),
-                              padding: const EdgeInsets.symmetric(vertical: 12),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-    );
-  }
-
-  Widget _buildStatCard(BuildContext context, String title, String value, IconData icon) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          children: [
-            Icon(icon, size: 32, color: Theme.of(context).colorScheme.primary),
-            const SizedBox(height: 8),
-            Text(
-              value,
-              style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            Text(
-              title,
-              style: Theme.of(context).textTheme.bodySmall,
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildMenuItem(BuildContext context, String title, IconData icon, VoidCallback onTap) {
-    return ListTile(
-      leading: Icon(icon),
-      title: Text(title),
-      trailing: const Icon(Icons.arrow_forward_ios),
-      onTap: onTap,
-    );
-  }
-}
-
-
