+++ lib/notes_screen.dart
@@ -1,0 +1,187 @@
+import 'package:flutter/material.dart';
+
+class NotesScreen extends StatefulWidget {
+  const NotesScreen({super.key});
+
+  @override
+  State createState() => _NotesScreenState();
+}
+
+class _NotesScreenState extends State {
+  String selectedCategory = 'All';
+  final List categories = ['All', 'Computer Science', 'Mathematics', 'Physics', 'Chemistry', 'Biology'];
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Notes'),
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.search),
+            onPressed: () {},
+          ),
+          IconButton(
+            icon: const Icon(Icons.filter_list),
+            onPressed: () {},
+          ),
+        ],
+      ),
+      body: Column(
+        children: [
+          // Category Filter
+          Container(
+            height: 60,
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: ListView.builder(
+              scrollDirection: Axis.horizontal,
+              itemCount: categories.length,
+              itemBuilder: (context, index) {
+                return Padding(
+                  padding: EdgeInsets.only(
+                    left: index == 0 ? 16 : 8,
+                    right: index == categories.length - 1 ? 16 : 0,
+                  ),
+                  child: FilterChip(
+                    label: Text(categories[index]),
+                    selected: selectedCategory == categories[index],
+                    onSelected: (selected) {
+                      setState(() {
+                        selectedCategory = categories[index];
+                      });
+                    },
+                  ),
+                );
+              },
+            ),
+          ),
+
+          // Notes List
+          Expanded(
+            child: ListView.builder(
+              padding: const EdgeInsets.all(16),
+              itemCount: 10,
+              itemBuilder: (context, index) {
+                return _buildNoteCard(context, index);
+              },
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildNoteCard(BuildContext context, int index) {
+    final notes = [
+      {
+        'title': 'Advanced Data Structures',
+        'subject': 'Computer Science',
+        'author': 'Sarah Wilson',
+        'rating': 4.8,
+        'downloads': 234,
+        'date': '2 days ago'
+      },
+      {
+        'title': 'Calculus II - Integration Techniques',
+        'subject': 'Mathematics',
+        'author': 'Robert Chen',
+        'rating': 4.6,
+        'downloads': 156,
+        'date': '5 days ago'
+      },
+      {
+        'title': 'Quantum Mechanics Basics',
+        'subject': 'Physics',
+        'author': 'Emily Davis',
+        'rating': 4.9,
+        'downloads': 89,
+        'date': '1 week ago'
+      },
+    ];
+
+    final note = notes[index % notes.length];
+
+    return Card(
+      margin: const EdgeInsets.only(bottom: 16),
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Row(
+              children: [
+                Expanded(
+                  child: Text(
+                    note['title'] as String,
+                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ),
+                IconButton(
+                  icon: const Icon(Icons.more_vert),
+                  onPressed: () {},
+                ),
+              ],
+            ),
+            const SizedBox(height: 8),
+            Row(
+              children: [
+                Chip(
+                  label: Text(
+                    note['subject'] as String,
+                    style: const TextStyle(fontSize: 12),
+                  ),
+                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
+                ),
+                const SizedBox(width: 8),
+                Text(
+                  'by ${note['author']}',
+                  style: Theme.of(context).textTheme.bodySmall,
+                ),
+              ],
+            ),
+            const SizedBox(height: 12),
+            Row(
+              children: [
+                Icon(Icons.star, size: 16, color: Colors.amber),
+                const SizedBox(width: 4),
+                Text('${note['rating']}'),
+                const SizedBox(width: 16),
+                Icon(Icons.download, size: 16, color: Colors.grey[600]),
+                const SizedBox(width: 4),
+                Text('${note['downloads']}'),
+                const Spacer(),
+                Text(
+                  note['date'] as String,
+                  style: Theme.of(context).textTheme.bodySmall,
+                ),
+              ],
+            ),
+            const SizedBox(height: 12),
+            Row(
+              children: [
+                Expanded(
+                  child: OutlinedButton.icon(
+                    onPressed: () {},
+                    icon: const Icon(Icons.visibility),
+                    label: const Text('Preview'),
+                  ),
+                ),
+                const SizedBox(width: 8),
+                Expanded(
+                  child: FilledButton.icon(
+                    onPressed: () {},
+                    icon: const Icon(Icons.download),
+                    label: const Text('Download'),
+                  ),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
